Accounts: Move PootleProfile.contributions to User.contributions
